requirements.txt
numpy==1.26.3
pandas==2.0.3
python-dotenv==1.0.1
requests==2.31.0